adds unit tests and fixes
