+++ scripts/google/covid_mobility/test1.csv
@@ -1,0 +1,4 @@
+﻿US,United States,Alabama,Baldwin County,5/24/20,,,,,,
+US,United States,Florida,Miami-Dade County,2/15/20,-1,-2,-3,,-4,
+US,United States,,,2/15/20,1,2,3,-4,-5,-6
+US,United States,Alabama,,2/15/20,1,2,3,4,5,6
